Changed directory structure, refactor. :1
