+++ Games/MatchByColor/MatchByColorGame.js
@@ -11,8 +11,8 @@ import reactMixin from 'react-mixin';
 import TimerMixin from 'react-timer-mixin';
 import randomstring from 'random-string';
 
-import AnimatedSprite from '../AnimatedSprite/AnimatedSprite';
-import HomeButton from '../HomeButton/HomeButton';
+import AnimatedSprite from '../../components/AnimatedSprite/AnimatedSprite';
+import HomeButton from '../../components/HomeButton/HomeButton';
 // props
 import leverSprite from '../../sprites/lever/leverCharacter';
 import signSprite from '../../sprites/sign/signCharacter';
